refactor(core): rename agenthub module to llm
