+++ packages/ema/src/channel/napcatqq_adapter.ts
@@ -1,5 +1,5 @@
 import type { ActorChatResponse } from "../actor";
-import type { InputContent, MIME } from "../agent_hub/schema";
+import type { InputContent, MIME } from "../llm/schema";
 import type {
   ChannelAdapter,
   ChannelAPICall,
