+++ packages/ema/src/config/index.ts
@@ -1,4 +1,4 @@
 export * from "./base";
 export * from "./global_config";
-export { resolveLLMModelDefinition } from "../agent_hub/models";
-export { RetryConfig } from "../agent_hub/retry";
+export { resolveLLMModelDefinition } from "../llm/models";
+export { RetryConfig } from "../llm/retry";
