+++ packages/ema/src/memory/base.ts
@@ -1,6 +1,6 @@
 import type { ActorChatInput, ActorChatResponse } from "../actor";
 import type { MessageReplyRef, SpeakerInformation } from "../channel";
-import type { InputContent } from "../agent_hub/schema";
+import type { InputContent } from "../llm/schema";
 
 export type BufferWriteMessage = ActorChatInput | ActorChatResponse;
