+++ packages/ema/src/shared/schema.ts
@@ -1,2 +1,2 @@
-export * from "../agent_hub/schema";
-export * from "../agent_hub/utils";
+export * from "../llm/schema";
+export * from "../llm/utils";
