+++ packages/ema/src/skills/query-chat-history-skill/index.ts
@@ -2,8 +2,8 @@ import { z } from "zod";
 import { Skill } from "../base";
 import type { ToolContext, ToolResult } from "../../tools/base";
 import type { ConversationMessageEntity } from "../../db/base";
-import type { ImageItem } from "../../agent_hub/schema";
-import { isImageMime } from "../../agent_hub/utils";
+import type { ImageItem } from "../../llm/schema";
+import { isImageMime } from "../../llm/utils";
 import { resolveSession } from "../../channel";
 import type { BufferMessage } from "../../memory/base";
 import { buildPromptFromBufferMessage } from "../../memory/utils";
